+++ src/rtc/rtc_manager.cpp
@@ -103,14 +103,24 @@ RTCManager::RTCManager(
                 resolve(cf.vp9_encoder, info.vp9_encoders),
                 resolve(cf.av1_encoder, info.av1_encoders),
                 resolve(cf.h264_encoder, info.h264_encoders), cf.simulcast,
-                cf.hardware_encoder_only));
+                cf.hardware_encoder_only
+#if defined(__linux__) && USE_NVCODEC_ENCODER
+                ,
+                cf.cuda_context
+#endif
+                ));
     media_dependencies.video_decoder_factory =
         std::unique_ptr<webrtc::VideoDecoderFactory>(
             absl::make_unique<MomoVideoDecoderFactory>(
                 resolve(cf.vp8_decoder, info.vp8_decoders),
                 resolve(cf.vp9_decoder, info.vp9_decoders),
                 resolve(cf.av1_decoder, info.av1_decoders),
-                resolve(cf.h264_decoder, info.h264_decoders)));
+                resolve(cf.h264_decoder, info.h264_decoders)
+#if USE_NVCODEC_ENCODER
+                    ,
+                cf.cuda_context
+#endif
+                ));
   }
 
   media_dependencies.audio_mixer = nullptr;