+++ src/sora/sora_client.h
@@ -45,6 +45,8 @@ struct SoraClientConfig {
   int data_channel_signaling_timeout = 180;
   boost::optional<bool> ignore_disconnect_websocket;
   int disconnect_wait_timeout = 5;
+  std::string client_cert;
+  std::string client_key;
 };
 
 class SoraClient : public std::enable_shared_from_this<SoraClient>,