+++ src/v4l2_video_capturer/v4l2_video_capturer.cpp
@@ -309,7 +309,7 @@ int32_t V4L2VideoCapturer::StartCapture(V4L2VideoCapturerConfig config) {
   // If driver doesn't support framerate control, need to hardcode.
   // Hardcoding the value based on the frame size.
   if (!driver_framerate_support) {
-    if (!_useNative && _currentWidth >= 800 &&
+    if (!config.use_native && _currentWidth >= 800 &&
         _captureVideoType != webrtc::VideoType::kMJPEG) {
       _currentFrameRate = 15;
     } else {
@@ -493,8 +493,8 @@ bool V4L2VideoCapturer::CaptureProcess() {
       // JPEG の先頭は SOI マーカー 0xffd8 で始まるのでチェックして落ちないようにする。
       if (_captureVideoType == webrtc::VideoType::kMJPEG && bytesused >= 2) {
         if (data[0] != 0xff || data[1] != 0xd8) {
-          RTC_LOG(LS_WARNING) << __FUNCTION__
-                              << " Invalid JPEG buffer frame skipped";
+          RTC_LOG(LS_WARNING)
+              << __FUNCTION__ << " Invalid JPEG buffer frame skipped";
         } else {
           unsigned int eosSearchSize = MJPEG_EOS_SEARCH_SIZE;
           uint8_t* p;